+++ beginner-gain-client/src/server/project.ts
@@ -12,3 +12,22 @@ export const makeProject = async ({ name, description, select, userId } : IMakeP
     });
     return response;
 };
+
+export const getProjects = async (userId: string) => {
+  if (!userId) return;
+  try {
+    const response = await axios.get(`${API_ENDPOINT}/project/user/${userId}`);
+    return response.data;
+  } catch (error) {
+    return error;
+  }
+}
+
+export const deleteProject = async (projectId: number) => {
+  try {
+    const response = await axios.delete(`${API_ENDPOINT}/project/${projectId}`);
+    return response.data;
+  } catch (error) {
+    return error;
+  }
+}
